refactor(hooks): extract chart lifecycle into useChart hook
Moves lightweight-charts init, resize observers, crosshair tooltips,
time scale sync, and data update logic out of GraphsPage into a focused
useChart hook. GraphsPage drops from 969 to 653 lines. Tooltip inline
styles replaced with CSS classes.
