+++ src/styles/graphs-page.css
@@ -269,6 +269,18 @@
   white-space: nowrap;
 }
 
+.graphs-tooltip-high {
+  color: rgb(34, 197, 94);
+}
+
+.graphs-tooltip-low {
+  color: rgb(239, 68, 68);
+}
+
+.graphs-tooltip-neutral {
+  color: rgb(148, 163, 184);
+}
+
 .graphs-info-badge {
   padding: 0.2rem 0.6rem;
   border-radius: 0.375rem;
